+++ src/utils/mod.rs
@@ -1,4 +1,5 @@
 pub mod datetime;
 pub mod html;
+pub mod html_parser;
 pub mod theme;
 pub mod theme_loader;